Separate lock.h: FreeRTOS, MQX, iOS
