+++ src/os/mqx/etcpal/os_signal.c
@@ -17,34 +17,20 @@
  * https://github.com/ETCLabs/EtcPal
  ******************************************************************************/
 
-#ifndef ETCPAL_OS_SEM_H_
-#define ETCPAL_OS_SEM_H_
+#include "etcpal/signal.h"
+#include "etcpal_mqx_common.h"
 
-#include <stdbool.h>
-#include <FreeRTOS.h>
-#include <semphr.h>
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-typedef SemaphoreHandle_t etcpal_sem_t;
-
-#define ETCPAL_SEM_HAS_TIMED_WAIT 1
-#define ETCPAL_SEM_HAS_POST_FROM_ISR 1
-#define ETCPAL_SEM_HAS_MAX_COUNT 1
-#define ETCPAL_SEM_MUST_BE_BALANCED 0
-
-bool etcpal_sem_create(etcpal_sem_t* id, unsigned int initial_count, unsigned int max_count);
-bool etcpal_sem_wait(etcpal_sem_t* id);
-bool etcpal_sem_try_wait(etcpal_sem_t* id);
-bool etcpal_sem_timed_wait(etcpal_sem_t* id, int timeout_ms);
-bool etcpal_sem_post(etcpal_sem_t* id);
-bool etcpal_sem_post_from_isr(etcpal_sem_t* id);
-void etcpal_sem_destroy(etcpal_sem_t* id);
-
-#ifdef __cplusplus
+bool etcpal_signal_timed_wait(etcpal_signal_t* id, int timeout_ms)
+{
+  if (timeout_ms < 0)
+  {
+    return (MQX_OK == _lwevent_wait_ticks(id, 1u, true, 0u));
+  }
+  else
+  {
+    MQX_TICK_STRUCT ticks_to_wait;
+    if (milliseconds_to_ticks(timeout_ms, &ticks_to_wait))
+      return (MQX_OK == _lwevent_wait_for(id, 1u, true, &ticks_to_wait));
+    return false;
+  }
 }
-#endif
-
-#endif /* ETCPAL_OS_SEM_H_ */
